requirements.txt
django
psycopg2-binary
gunicorn
whitenoise
dj-database-url
Pillow
django-storages
python-dotenv
folium
jupyterlab
notebook
civilpy
honcho
google-generativeai
Markdown
